+++ changelog.md
@@ -10,6 +10,7 @@
 - Changed `example/test-request` to `Sandbox` instead of `Production`
 - Use `Network.HTTP.Types.encodePathSegments` when making `RequestPath` in authenticated requests
 - Support exchange limits endpoint
+- Added `min_market_funds` and `max_market_funds` to `Product`
 
 # Version 0.8.2.0